First stab at cosmetic filtering
First stab at cosmetic filtering with the python-adblock library.

Cosmetic filtering is done in two steps:
1. Lookup url-specific cosmetic selectors.
2. Lookup generic selectors based on the element classes and ids present in the page.

In our code, we perform step 1 in a load_started hook. The results are then translated
to javascript and inserted into the page using Tab.run_js_async. The injected javascript
also returns all the class names and ids that appear in the page. A callback then takes
this information and runs step 2.

Note that scriptlet injection is not yet available because extra resources need to be
downloaded and added to the engine before script injection resolves properly.