First stab at scriptlet injection
To add resources to the engine, run :adblock-update-resources. Then, the resources
should automatically download, cache, and add themselves to the engine.

The way the resources are determined is similar to rust Brave's adblocker code.
They download uBlock's redirect-engine.js, parse it, and select resources based on that.
They also parse and add the functions and function templates from uBlock's
scriptlets.js. We do basically the same thing.

Note that this version requires the version of python-adblock whose add_resource
function for the engine has 4 arguments (including aliases). This will likely be version
0.5.3.

Scriptlet injection seems to mostly work, but there are still some annoying caveats.
Because currently scriptlets are injected after the page is loaded, ads on youtube will
will not be blocked if the video is opened a new tab.